Fixed broken secondary sights on gamepads
Fixed a bug where the mod would cause users on controllers or gamepads to be unable to toggle their weapons' sight gadgets (such as 45deg sights or magnifiers).
Changed how controller binds are stored- now they are change-able in the mod settings (though this feature is not officially supported, as those bindings' default actions are not appropriately disabled in PlayerStandard. Maybe in the future.)
Version bump to 3.4
Minor misc cleaning